move signalToFeatureCollection to mapHelpers
